refact(project): add connection to struct
This was a lot more complicated than I expected, because you cannot share a
rusqlite `Connection` object across unwrap boundaries (you can share it across
threads, but our `Async` does automatic unwrapping …).

So what I do here is that whenever we want to clone a `Project`, I create
something I call “skeleton”, which is a clone of the project without the
connection object (just the sqlite file path), then on the other side of the
boundary I turn the skeleton back into a Project by connecting to the database
again. I think this is quite an okay solution.